Refactor AudioManager, add NowPlayingManager
